Fix Firebase wiring and reorganize project structure
