+++ config/scripts.js
