+++ package.json
@@ -3,11 +3,12 @@
   "private": true,
   "type": "module",
   "scripts": {
-    "test": "node --test",
-    "lint": "eslint .",
+    "test": "node --test tests/*.test.js",
+    "lint": "eslint . --config config/eslint.config.js",
     "format": "prettier --check .",
     "format:write": "prettier --write .",
-    "dev": "vite"
+    "dev": "vite --config config/vite.config.js",
+    "build": "vite build --config config/vite.config.js"
   },
   "devDependencies": {
     "eslint": "^9.20.1",
