+++ public/index.html
@@ -4,7 +4,7 @@
   <meta charset="UTF-8">
   <title>Life RPG Dashboard</title>
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <link rel="stylesheet" href="style.css">
+  <link rel="stylesheet" href="./style.css">
 </head>
 
 <body>
@@ -85,6 +85,6 @@
     </section>
   </div>
 
-  <script type="module" src="ui.js"></script>
+  <script type="module" src="../src/ui/ui.js"></script>
 </body>
 </html>
