+++ src/core/financeLogic.js
