+++ src/core/firebaseService.js
