+++ src/core/rewardLogic.js
