+++ src/core/validation.js
